Move LocalizationExt Trait to `localization.rs`
It was previously with the metadata. I couldn't find it when I went
looking for it so I think that was a bad spot.
